requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==22.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cvxopt==1.3.2
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
discrete-signals==0.7.3
dnspython==2.6.1
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
func_timeout==4.3.5
funcy==2.0
gensim==4.3.2
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.2.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.7
language-data==1.1
lark==1.1.9
lenses==0.5.0
lxml==5.2.2
marisa-trie==0.7.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.2
matplotlib-inline==0.1.7
mdurl==0.1.2
metric-temporal-logic==0.4.0
mistune==3.0.2
mlxtend==0.21.0
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.24.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parsimonious==0.8.1
parso==0.8.4
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pm4py==2.5.1
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.2
pydantic-mongo==2.3.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydotplus==2.0.2
Pygments==2.18.0
pylogics==0.2.1
pymongo==4.8.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.12.0
seaborn==0.12.1
semconstmining @ git+https://github.com/a-rebmann/semantic-constraint-miner.git@f2840dd1896fe1f9f4518a9a44bde9bb3c99c492
Send2Trash==1.8.3
sentence-transformers==3.0.1
shellingham==1.5.4
singledispatch==4.1.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
stringcase==1.2.0
StringDist==1.0.9
sympy==1.13.0
terminado==0.18.1
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typer==0.4.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.20.0
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.11
wrapt==1.16.0
